fixed bug. The function exponentiateWithPreComputedValues gets an exponent of type BigInteger, not int.
git-svn-id: https://svn.cs.biu.ac.il/svn/development/SDK@1044 ba1fec17-5bc3-4c1d-a110-6a66ae20c876

+++ Code/JavaSrc/edu/biu/scapi/primitives/dlog/DlogGroup.java
@@ -184,7 +184,7 @@ public interface DlogGroup {
 	 * @return the exponentiation result
 	 * @throws UnInitializedException 
 	 */
-	public GroupElement exponentiateWithPreComputedValues(GroupElement groupElement, int exponent) throws UnInitializedException;
+	public GroupElement exponentiateWithPreComputedValues(GroupElement groupElement, BigInteger exponent) throws UnInitializedException;
 	
 	/**
 	 * Converts a byte array to a GroupElement.
